BUG: Get the word and reading in a more robust way.
I attempted to look up a word found in a title of a song, and jisho.org had an
entry which treated the word as a proper noun and just had a wikipedia sense.
This word used a long alphanumeric string as the slug, not the search word.
Also, there was no reading information. There was also no is_common key.[0]

This change changes how we find some information:

word now comes from: data[0].japanese[0].word then data[0].slug
is_common now defaults to false
reading defaults to empty string, which means no known reading.

[0]: For future reference, and maybe an interesting test case, the keyword
provided was カリソメ, in reference to the song カリソメ乙女 by 椎名 林檎(しいな
りんご).